+++ tests/test_pdfalyzer.py
@@ -21,7 +21,7 @@ def test_help_option():
 
 def test_pdfalyzer_basic_tree(adobe_type1_fonts_pdf_path, analyzing_malicious_documents_pdf_path):
     type1_tree = _run_with_args(adobe_type1_fonts_pdf_path, '-t')
-    _assert_line_count_within_range(88, type1_tree)
+    _assert_line_count_within_range(90, type1_tree)
     analyzing_malicious_tree = _run_with_args(analyzing_malicious_documents_pdf_path, '-t')
     _assert_line_count_within_range(1004, analyzing_malicious_tree)
 
@@ -52,7 +52,7 @@ def _run_with_args(pdf, *args) -> str:
 def _assert_line_count_within_range(line_count, text):
     lines_in_text = len(text.split("\n"))
 
-    if not isclose(line_count, lines_in_text, rel_tol=0.02):
+    if not isclose(line_count, lines_in_text, rel_tol=0.05):
         for i, line in enumerate(text.split("\n")):
             print(f"{i}: {line}")